Hide replays for unplayed matches
this is a bug. The workers can fail to set the match result, but succeed
in uploading the replay. This ofc makes not much sense to the user, so
we hide the replay until the match gets replayed automatically.

A better fix would be to not accept the replay if it is flagged as not
ran