Handle agent ratings object not existing for old matches
This happens when there are no colosseum workers running, which causes
matches to be pending on old seasons. When the match eventually runs,
there will be no agent rating for that agent / season pair and it will
break. This PR attempts to fix it by hacking it so if it breaks, it
unbreaks itself. A proper fix would "just" ensure that there is always
an agent rating for all agent / season pair. An idea is having a service
that does that, and it is called when either a season or an agent are
created.